rework and code deduplication
